pyproject.toml
[project]
name = "desktop-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "pandas>=2.3.0",
    "pyautogui>=0.9.54",
    "pyscreeze>=1.0.1",
    "transformers>=4.53.0",
]
